Revert "Upgraded to version 12.0.0."
This reverts commit 3fc0432c3af875739dcc9bb38c190952b7348f96.

I mistakenly copied the `clang` folder under `utils` rather than `tools`!

+++ include/llvm/CodeGen/DwarfStringPoolEntry.h
@@ -21,7 +21,7 @@ struct DwarfStringPoolEntry {
   static constexpr unsigned NotIndexed = -1;
 
   MCSymbol *Symbol;
-  uint64_t Offset;
+  unsigned Offset;
   unsigned Index;
 
   bool isIndexed() const { return Index != NotIndexed; }
@@ -47,7 +47,7 @@ public:
     assert(getMapEntry()->second.Symbol && "No symbol available!");
     return getMapEntry()->second.Symbol;
   }
-  uint64_t getOffset() const { return getMapEntry()->second.Offset; }
+  unsigned getOffset() const { return getMapEntry()->second.Offset; }
   bool isIndexed() const { return MapEntryAndIndexed.getInt(); }
   unsigned getIndex() const {
     assert(isIndexed());
